fix: don't panic when given a bad topology selector
