+++ pkg/operator/scheme/scheme.go
@@ -16,8 +16,9 @@ package scheme
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+
+	"github.com/samber/lo"
 
 	"github.com/aws/karpenter-core/pkg/apis"
 )
@@ -27,6 +28,6 @@ var (
 )
 
 func init() {
-	utilruntime.Must(clientgoscheme.AddToScheme(Scheme))
-	utilruntime.Must(apis.AddToScheme(Scheme))
+	lo.Must0(clientgoscheme.AddToScheme(Scheme))
+	lo.Must0(apis.AddToScheme(Scheme))
 }
